+++ lives-plugins/plugins/playback/video/vloopback2.c
@@ -284,7 +284,7 @@ static void make_path(const char *fname, int pid, const char *ext) {
 }
 
 
-boolean init_screen (int width, int height, boolean fullscreen, uint32_t window_id, int argc, char **argv) {
+boolean init_screen (int width, int height, boolean fullscreen, uint64_t window_id, int argc, char **argv) {
   int i=0,idx=0,ret_code;
   int afd,audio;
